fix: Medication CRUD 수정

+++ src/main/java/com/solution/Ongi/domain/medication/service/MedicationService.java
@@ -2,17 +2,20 @@ package com.solution.Ongi.domain.medication.service;
 
 import com.solution.Ongi.domain.medication.Medication;
 import com.solution.Ongi.domain.medication.dto.CreateMedicationRequest;
+import com.solution.Ongi.domain.medication.dto.MedicationDTO;
+import com.solution.Ongi.domain.medication.dto.UpdateMedicationRequest;
 import com.solution.Ongi.domain.medication.repository.MedicationRepository;
 import com.solution.Ongi.domain.user.User;
 import com.solution.Ongi.domain.user.repository.UserRepository;
 import com.solution.Ongi.domain.user.service.UserService;
+import com.solution.Ongi.global.response.code.ErrorStatus;
+import com.solution.Ongi.global.response.exception.GeneralException;
 import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 @Service
 @Transactional
@@ -41,16 +44,44 @@ public class MedicationService {
         return medicationRepository.save(medication);
     }
 
-    //유저의 Meal 전체 조회
-    public List<Medication> getAllMedication(Long userId){
-        userService.getUserByIdOrThrow(userId);
-        return medicationRepository.findByUserId(userId);
+    //유저의 Medication 전체 조회
+    public List<MedicationDTO> getAllMedication(String loginId){
+        User user = userService.getUserByLoginIdOrThrow(loginId);
+        List<Medication> medications = medicationRepository.findAllByUserId(user.getId());
+        List<MedicationDTO> result = medications.stream()
+            .map(MedicationDTO::new)
+            .toList();
+        return result;
     }
 
-    //Meal 삭제
-    public void deleteMedication(Long medication_id){
+    // Medication 삭제
+    public void deleteMedication(String loginId, Long medication_id){
+        User user = userService.getUserByLoginIdOrThrow(loginId);
         Medication medication=medicationRepository.findById(medication_id)
-                .orElseThrow(()->new RuntimeException("약이 존재하지 않습니다."));
+                .orElseThrow(()->new GeneralException(ErrorStatus.MEDICATION_NOT_FOUND));
+
+        if (!medication.getUser().getId().equals(user.getId())){
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED_ACCESS);
+        }
+
         medicationRepository.delete(medication);
     }
+
+    public void updateMedication(String loginId, Long medicationId, UpdateMedicationRequest request) {
+        User user = userService.getUserByLoginIdOrThrow(loginId);
+        Medication medication = medicationRepository.findById(medicationId)
+            .orElseThrow(() -> new GeneralException(ErrorStatus.MEDICATION_NOT_FOUND));
+
+        if (!medication.getUser().getId().equals(user.getId())) {
+            throw new GeneralException(ErrorStatus.UNAUTHORIZED_ACCESS);
+        }
+
+        List<LocalTime> convertedTimeList = request.timeList().stream()
+            .map(LocalTime::parse) // "12:30" → LocalTime
+            .toList();
+
+        medication.update(request.title(), convertedTimeList);
+    }
+
+    // Medication 수정
 }
